Make (un)association stages work on primary keys
The pipeline concept is based on:

1. Identifying existing content units that represent content that should be
  part of the new repo version (QueryExistingContentUnits stage)

2. Saving all other new content units (ContentUnitSaver stage)

In conclusion, all declarative_content.content instances _must_ be in the
database before they enter the (un)association stages.  All the heavy
lifting based on unit keys has already been done.

Thus, the (un)association stages should work on primary keys only.

This also enables to use pipelines which are "local" to the repo version,
i.e. pipelines that do not rely on global natural keys for content, but
only query and associate existing content in the base repository version.

[noissue]