Problem: plugin API missing pull-through cache interfaces
Solution: add interfaces that enable pull-through cache for plugins

This patch extends the Remote API. It adds an interface for determining the full URL of a RemoteArtifact
by examining the relative path and the Remote's url. It also adds an interface for determining the type of
content that should be available at a relative path of a remote. These interfaces are used by the content
app when serving content in the pull-through cache mode. The content app needs to determine what URL is
actually being requested and what kind of content is being requested by the client.

This patch also extends the Content API. It adds a new constructor method to the Content model. This new
interface is a constructor that take an Artifact and a relative path to produce an unsaved instance of a
specific content type.

Plugin writers are expected to provide specific implementations of these methods.

re: #3894
https://pulp.plan.io/issues/3894