Problem: downloader factory requires ssl certs as files
Solutions: write client cert and key to temporary file

This patch enables the download factory to construct downloaders from certificates stored in the db.
The cadata object is used to pass in the CA for verifying the server[0]. The client key and client cert
are written to temporary files just before being loaded into the sslcontext. The temporary files are
then destroyed.

[0] https://docs.python.org/3.5/library/ssl.html#ssl-certificates

re: #4506
https://pulp.plan.io/issues/4506